move hyperlane constants to src/common/constants
