+++ src/common/Constants.ts
@@ -9,6 +9,8 @@ import {
   ZERO_ADDRESS,
   bnUint32Max,
   EvmAddress,
+  toWei,
+  BigNumber,
 } from "../utils";
 import {
   BaseBridgeAdapter,
@@ -948,4 +950,38 @@ export const ARBITRUM_ORBIT_L1L2_MESSAGE_FEE_DATA: {
     feePayer: "0x0d57392895Db5aF3280e9223323e20F3951E81B1", // DonationBox
     feeToken: TOKEN_SYMBOLS_MAP.AZERO.addresses[CHAIN_IDs.MAINNET],
   },
+};
+
+// source: https://github.com/hyperlane-xyz/hyperlane-registry/blob/346b18c4314cf96b41ae2da781f58fb832dbe1f8/deployments/warp_routes/EZETH/arbitrum-base-berachain-blast-bsc-ethereum-fraxtal-linea-mode-optimism-sei-swell-taiko-unichain-worldchain-zircuit-config.yaml
+export const HYPERLANE_ROUTERS: { [chainId: number]: { [tokenAddress: string]: string } } = {
+  [CHAIN_IDs.MAINNET]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.MAINNET].toLowerCase()]: "0xC59336D8edDa9722B4f1Ec104007191Ec16f7087",
+  },
+  [CHAIN_IDs.ARBITRUM]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.ARBITRUM].toLowerCase()]: "0xB26bBfC6d1F469C821Ea25099017862e7368F4E8",
+  },
+  [CHAIN_IDs.BASE]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.BASE].toLowerCase()]: "0x2552516453368e42705D791F674b312b8b87CD9e",
+  },
+  [CHAIN_IDs.BLAST]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.BLAST].toLowerCase()]: "0x486b39378f99f073A3043C6Aabe8666876A8F3C5",
+  },
+  [CHAIN_IDs.MODE]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.MODE].toLowerCase()]: "0xC59336D8edDa9722B4f1Ec104007191Ec16f7087",
+  },
+  [CHAIN_IDs.LINEA]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.LINEA].toLowerCase()]: "0xC59336D8edDa9722B4f1Ec104007191Ec16f7087",
+  },
+  [CHAIN_IDs.UNICHAIN]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.UNICHAIN].toLowerCase()]: "0xFf0247f72b0d7ceD319D8457dD30622a2bed78B5",
+  },
+  [CHAIN_IDs.OPTIMISM]: {
+    [TOKEN_SYMBOLS_MAP.ezETH.addresses[CHAIN_IDs.OPTIMISM].toLowerCase()]: "0xacEB607CdF59EB8022Cc0699eEF3eCF246d149e2",
+  },
+};
+
+// 0.1 ETH is a default cap for chains that use ETH as their gas token
+export const HYPERLANE_DEFAULT_FEE_CAP = toWei("0.1");
+export const HYPERLANE_FEE_CAP_OVERRIDES: { [chainId: number]: BigNumber } = {
+  // all supported chains that have non-eth for gas token should go here. Example: BSC
 };
